+++ src/routers/task_result.py
@@ -1,4 +1,4 @@
-from parser.task import get_tasks_data
+from parser.task_results import get_tasks_results_data
 from typing import List
 
 from fastapi import APIRouter, Depends, HTTPException
@@ -6,17 +6,17 @@ from sqlmodel import Session, select
 
 from database import get_db
 from models.competition import CompetitionModel
-from models.task import TaskModel
+from models.task_result import TaskResultModel
 
 router = APIRouter()
 
 
 @router.get(
-    "/get_tasks_for_competition", summary="Add a new competition to the scraper"
+    "/get_task_results_for_competition", summary="Add a new competition to the scraper"
 )
-async def get_tasks_for_competition(
+async def get_task_results_for_competition(
     competition_id: int, session: Session = Depends(get_db)
-) -> List[TaskModel]:
+) -> List[TaskResultModel]:
     try:
         result = session.exec(
             select(CompetitionModel).where(
@@ -24,7 +24,7 @@ async def get_tasks_for_competition(
             )
         ).all()
         if result:
-            return get_tasks_data(result[0])
+            return get_tasks_results_data(result[0])
         return []
     except Exception as ex:
         raise HTTPException(status_code=500, detail=f"{ex}") from ex