+++ src/tests/test_competitor.py
@@ -3,17 +3,21 @@ from pytest_mock import MockerFixture
 
 from tests.conftest import (
     ONE_COMPETITION_DUMMY_DATA,
-    get_expected_results_from_file,
+    get_expected_competitor_and_task,
     get_test_files,
     get_xml_tree_from_file,
 )
 
 TESTS_TO_RUN: str = "**/*.html"
+DATA_TESTS_FOLDER: str = "data/competitors_and_tasks"
 
 
 @pytest.mark.parametrize(
     "user_data_to_add, html_file",
-    [(ONE_COMPETITION_DUMMY_DATA, x) for x in get_test_files(TESTS_TO_RUN)],
+    [
+        (ONE_COMPETITION_DUMMY_DATA, x)
+        for x in get_test_files(DATA_TESTS_FOLDER, TESTS_TO_RUN)
+    ],
 )
 def test_competitor_parser(
     test_client, user_data_to_add, html_file, mocker: MockerFixture
@@ -28,13 +32,10 @@ def test_competitor_parser(
     response = test_client.get(
         "/competitor/get_competitors_in_competition", params={"competition_id": 1}
     )
-    expected = get_expected_results_from_file(html_file)
-    assert response.status_code == expected["expected_response"]
+    expected = get_expected_competitor_and_task(html_file)
+    assert response.status_code == expected.expected_response
     if response.status_code == 200:
-        assert len(response.json()) == expected["expected_number_competitors"]
-        # assert [
-        #     x["name"] for x in sorted(response.json(), key=lambda x: x["task_id"])
-        # ] == expected["tasks"]
+        assert len(response.json()) == expected.expected_number_competitors
 
 
 def test_competitor_empty_competitions(test_client):